Splitting vibration body to add additional screens
